+++ src/main/java/com/vsii/dao/ISupplierDAO.java
@@ -6,7 +6,7 @@ import com.vsii.entity.Supplier;
 
 public interface ISupplierDAO {
 	public List<Supplier> getAllSuppliers();
-	public Supplier getSupplierById(int supplierId);
+	public Supplier getSupplierById(long supplierId);
     public void addSupplier(Supplier supplier);
     public void updateSupplier(Supplier supplier);
     public void deleteSupplier(int supplierId);
